Refactor and fix run_await logic
Before this change, run_await was awaiting for another run_await, and
that run_await was waiting for another run_await, and so on. Therefore,
the code after gathering the tasks was never being executed.

I refactored the code and now it is a traditional (async) infinite loop.
It creates a different task for each websocket message, so multiple
processes could be running at once.